Add Powerline Evil Colors
Powerline Evil is an emacs package that adds an indicator to the Emacs
modeline. The color of the indicator changes depending on which VI mode
emacs is in (green for Normal mode, Orange for Visual, etc). This commit
adds settings for these indicators to dracula-theme.el, so that they
use the Dracula versions of green, blue, orange, etc. instead of the
generic colors used by powerline-evil.el.