Do not advertize dracula-use-24-bit-colors-on-256-colors-terms in README
The issue 59 proves again that 256 colors terminal cannot handle the
24bit colors of the graphical definition.  Thus regular user should not
try to activate this new setting, or, if they want to, must fully
aknowledge the possible drawbacks.

I also add a long comment to this setting to explain one of the possible
problems and a solution to it.

See: https://github.com/dracula/emacs/issues/59